refactor: use int32_t to save cf idx

+++ src/storage/include/storage/binlog.h
@@ -13,7 +13,7 @@ enum class OperateType { kNoOperate = 0, kPut, kDelete };
 enum DataType { kAll, kStrings, kHashes, kLists, kZSets, kSets };
 
 struct BinlogEntry {
-  int8_t cf_idx_;
+  int32_t cf_idx_;
   OperateType op_type_;
   Slice key_;
   std::optional<Slice> value_;
@@ -23,14 +23,14 @@ class Binlog {
  public:
   explicit Binlog(DataType type) : data_type_(type) {}
 
-  // void AppendOperation(int8_t cfid, OperateType type, const Slice& key,
+  // void AppendOperation(int32_t cfid, OperateType type, const Slice& key,
   //                      const std::optional<Slice>& value = std::nullopt) {
   //   entries_.emplace_back(cfid, type, key, value);
   // }
-  void AppendPutOperation(int8_t cfid, const Slice& key, const Slice& value) {
+  void AppendPutOperation(int32_t cfid, const Slice& key, const Slice& value) {
     entries_.emplace_back(cfid, OperateType::kPut, key, value);
   }
-  void AppendDeleteOperation(int8_t cfid, const Slice& key) {
+  void AppendDeleteOperation(int32_t cfid, const Slice& key) {
     entries_.emplace_back(cfid, OperateType::kDelete, key, std::nullopt);
   }
